restructuring test suite
